Always clear regardless of color/depth value
That was a stupid one... didn't realize OpenGL always you to first
specify the color and depth values, then do `glClear` to actually clear.

`GLWrapper` verison of this would only change the color/depth value when
necessary, but does `glClear` regardless.

Here we need to do the clear functions regardless, as `ClearColorTarget`
acts as a `glClearColor` followed by `glClear`, and `ClearDepthStencil` acts as a
`glClearDepthf` followed by `glClear`.