Add proof-of-concept `VeldridGraphicsBackend` reimplementation
`VeldridGraphicsBackend` has now been reimplemented as a low-level
`IGraphicsBackend` implementation, with the introduction of a `Resource`
object that all device resources (such as `RendererTexture`,
`VertexBuffer`, `Shader`) will have, in order for
`VeldridGraphicsBackend` to grab a handle to the veldrid resources for
use.

Note that this is still not complete and will fail building.