Always convert to sRGB space on backbuffer
Now with Veldrid (including its OpenGL), it appears the sRGB conversion process needs to always be applied when drawing to the backbuffer (i.e. swapchain framebuffer), not only on OpenGL ES.

I'm not really sure why this is the case, but since this is working
correctly I think there's no need to dig too deep into the OpenGL rabbit
hole to find out what's going on.

We could also go down the path of using `SwapchainSrgbFormat = true`
(Veldrid's way of enabling `GL_FRAMEBUFFER_SRGB`) and call it a day, but
I'll leave that to be decided in
https://github.com/ppy/osu-framework/issues/2790.