Allow framework consumers to override `FrameworkSetting` ranges
To support overriding the range of window size. Could potentially be used
for more.

The constraint existed years ago when the framework was powered using OpenTK/osuTK, but since the switch to SDL, the constraint was left unimplemented.

Since it has been over two years ago while the constraint is ineffective, this is considered a "breaking change" for the framework consumers.

In order to give the consumer enough flexibility, we introduced a method overridable in `Game` which can be used to override the constraint.
```diff
+        protected override IDictionary<FrameworkSetting, (object, object)> GetFrameworkConfigRanges()
+            => new Dictionary<FrameworkSetting, (object, object)>
+            {
+                { FrameworkSetting.WindowedSize, (new Size(250, 200), new Size(1000, 800)) },
+            };
```